bridge: try harder to ensure Endpoints exited
When the bridge shuts down in an orderly fashion from an EOF on stdin,
we signal the endpoints to exit, wait for them to do so, and only close
the connection after that's done.

When the connection closes suddenly, such as from a BrokenPipeError, we
don't do this waiting around.  The result is that the endpoints are
never signalled to clean themselves up, and (crucially) we don't wait
for them to do so.

Let's move the communicate() method from the protocol level to the
router so that we can make it aware of endpoints.  Then we make two
changes:

 - if we didn't already call eof_received() then make sure it happens on
   disconnect of the transport.  This is what is responsible for asking
   the endpoints to close.

 - from the end of communicate(), unconditionally wait for all endpoints
   to have closed down.  This way we'll never get endpoints still
   running after the event loop has closed.

We accomplish that by adding a bit of redundant state: an asyncio.Event
which can be awaited, and tracks if we have no endpoints left.

We also drop calling .shutdown() on all of the routing rules.  This is
no longer necessary, as we're signalling all of the endpoints directly.

A minor tweak is needed to a test which was using communicate() on a
subclass of CockpitProtocolServer — it can derive from Router now.

This should remove all of the warnings we've been seeing about "Event
loop is closed" and "Task was destroyed but it is pending".

Closes #18355
